// src/app/interfaces/data.ts
export interface CourseListItem {
  ownerEntityId: string;
  programCourseKey: string;
  forEach: any;
    courseCode: string;
    ownerEntityName: string;
    program: string;
    branch: string;
    specialisation: string;
    semester: string;
    A: string;
    Am: string;
    B: string;
    Bm: string;
    C: string;
    CM: string;
   D: string;
    Dm: string;
    E: string;
    Em: string;
    F: string;
    resultStatus: string;
    awardSheetStatus: string;
  }
  